+++ lib/domain/env_config.dart
@@ -17,19 +17,18 @@ class EnvConfig extends Equatable {
   final DateTime packTime;
   final EnvSign envSign;
 
-  const EnvConfig(
-    this.appName,
-    this.libVersion,
-    this.packTime,
-    this.envSign,
-  );
+  const EnvConfig(this.appName,
+      this.libVersion,
+      this.packTime,
+      this.envSign,);
 
   /// 使用本构造前, 请确保你不是新手
   ///
   /// 本构造适用于单独配置某一个仅使用EnvSign的GetArchPackage,
   /// 当然, 如果你的整个项目中都没有用到其他属性,也可以在globalConfig中使用本构造
   const EnvConfig.sign(this.envSign)
-      : this.appName = null,
+      : assert(envSign != null),
+        this.appName = null,
         this.libVersion = null,
         this.packTime = null;
 
@@ -39,25 +38,20 @@ class EnvConfig extends Equatable {
   final bool stringify = true;
 }
 
-class EnvSign {
-  final String _sign;
-  const EnvSign.fromString(this._sign);
+enum EnvSign {
+  dev,
+  test,
+  staging,
+  prod,
+}
 
-  static const EnvSign dev = const EnvSign.fromString('dev');
-  static const EnvSign test = const EnvSign.fromString('test');
-  static const EnvSign prod = const EnvSign.fromString('prod');
+extension EnvSignX on Iterable<EnvSign> {
+  static const _map = {
+    'dev': EnvSign.dev,
+    'test': EnvSign.test,
+    'staging': EnvSign.staging,
+    'prod': EnvSign.prod,
+  };
 
-  @override
-  String toString() => _sign;
-
-  @override
-  bool operator ==(Object other) =>
-      (identical(this, other)) ||
-      (other is String && other == _sign) ||
-      (other is EnvSign &&
-          runtimeType == other.runtimeType &&
-          _sign == other._sign);
-
-  @override
-  int get hashCode => _sign.hashCode;
+  fromString(String sign) => _map[sign];
 }