+++ lib/domain/i_failure.dart
@@ -16,7 +16,8 @@ abstract class Failure extends Equatable {
   }
 
   /// 自动包装
-  factory Failure.auto(dynamic e) => e is Failure ? e : UnknownFailure(e);
+  factory Failure.auto(dynamic e, [dynamic stackTrace, String msg]) =>
+      e is Failure ? e : UnknownFailure('$msg[$e]', stackTrace);
 
   @override
   List<Object> get props => [this.reportFailureType, this.msg];