+++ lib/gac_base.dart
@@ -1,0 +1,6 @@
+export 'package:get_it/get_it.dart';
+export 'package:equatable/equatable.dart';
+export 'package:dartz/dartz.dart'
+    show Either, Unit, left, right, Left, Right, Tuple2, Tuple3;
+export 'package:rxdart/rxdart.dart';
+export 'package:meta/meta.dart';