+++ test/get_arch_application_test.dart
@@ -19,6 +19,6 @@ main() {
   });
 
   test('get EnvSign from String ', () {
-    expect(EnvSign.fromString('dev'), EnvSign.dev);
+    expect(EnvSign.values.fromString('dev'), EnvSign.dev);
   });
 }